+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improvement: Implement `page` param in requests;
 - Improvement: Optimize selecting fields from database on `GET`;
+- Improvement: Replace all error responses in blueprints with `negotiate`;
 - Fix: Issue with custom `id` on `create` resource;
 - Typo: Change `serverError` to `notFound` in `update` blueprint;